# services/api/openapi.yaml
openapi: 3.1.0
info:
  title: Rune API
  version: 0.0.0
paths:
  /auth/login:
    post:
      tags:
      - Authentication
      summary: User login
      description: Authenticate user with email and password. Returns access and refresh
        tokens, and sets an HTTP-only cookie.
      operationId: login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TokenResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/refresh:
    post:
      tags:
      - Authentication
      summary: Refresh access token
      description: Generate a new access token using a valid refresh token. The refresh
        token remains unchanged.
      operationId: refresh_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TokenResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/logout:
    post:
      tags:
      - Authentication
      summary: User logout
      description: Logout the current user by revoking their refresh token and clearing
        the authentication cookie.
      operationId: logout_auth_logout_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_NoneType_'
      security:
      - OAuth2PasswordBearer: []
...
  /auth/first-admin-signup:
    post:
      tags:
      - Authentication
      summary: First-time admin signup
      description: Create the first admin account. Only available when no users exist
        in the system. Includes race condition protection.
      operationId: first_admin_signup_auth_first_admin_signup_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FirstAdminSignupRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_FirstAdminSignupResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /templates/:
    get:
      tags:
      - Templates
      summary: List Templates
      description: Get all templates accessible to the current user (public + their
        own).
      operationId: list_templates_templates__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_List_TemplateSummary__'
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - Templates
      summary: Create Template
      description: Create a new template.
      operationId: create_template_templates__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TemplateDetail_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /templates/{template_id}:
    get:
      tags:
      - Templates
      summary: Get Template
      description: Get a specific template by ID.
      operationId: get_template_templates__template_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: integer
          title: Template Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TemplateDetail_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Templates
      summary: Delete Template
      description: Delete a template.
      operationId: delete_template_templates__template_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: integer
          title: Template Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /templates/{template_id}/use:
    post:
      tags:
      - Templates
      summary: Use Template
      description: Mark a template as used (increment usage count) and return its
        workflow data.
      operationId: use_template_templates__template_id__use_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: integer
          title: Template Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TemplateWorkflowData_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiResponse_FirstAdminSignupResponse_:
      properties:
        success:
          type: boolean
          title: Success
          description: Whether the request was successful
          default: true
        message:
          type: string
          title: Message
          description: Human-readable message
          default: Success
        data:
          $ref: '#/components/schemas/FirstAdminSignupResponse'
          description: Response data
      type: object
      required:
      - data
      title: ApiResponse[FirstAdminSignupResponse]
...
    ApiResponse_List_TemplateSummary__:
      properties:
        success:
          type: boolean
          title: Success
          description: Whether the request was successful
          default: true
        message:
          type: string
          title: Message
          description: Human-readable message
          default: Success
        data:
          items:
            $ref: '#/components/schemas/TemplateSummary'
          type: array
          title: Data
          description: Response data
      type: object
      required:
      - data
      title: ApiResponse[List[TemplateSummary]]
    ApiResponse_NoneType_:
      properties:
        success:
          type: boolean
          title: Success
          description: Whether the request was successful
          default: true
        message:
          type: string
          title: Message
          description: Human-readable message
          default: Success
        data:
          type: 'null'
          title: Data
          description: Response data
      type: object
      required:
      - data
      title: ApiResponse[NoneType]
    ApiResponse_TemplateDetail_:
      properties:
        success:
          type: boolean
          title: Success
          description: Whether the request was successful
          default: true
        message:
          type: string
          title: Message
          description: Human-readable message
          default: Success
        data:
          $ref: '#/components/schemas/TemplateDetail'
          description: Response data
      type: object
      required:
      - data
      title: ApiResponse[TemplateDetail]
    ApiResponse_TemplateWorkflowData_:
      properties:
        success:
          type: boolean
          title: Success
          description: Whether the request was successful
          default: true
        message:
          type: string
          title: Message
          description: Human-readable message
          default: Success
        data:
          $ref: '#/components/schemas/TemplateWorkflowData'
          description: Response data
      type: object
      required:
      - data
      title: ApiResponse[TemplateWorkflowData]
    ApiResponse_TokenResponse_:
      properties:
        success:
          type: boolean
          title: Success
          description: Whether the request was successful
          default: true
        message:
          type: string
          title: Message
          description: Human-readable message
          default: Success
        data:
          $ref: '#/components/schemas/TokenResponse'
          description: Response data
      type: object
      required:
      - data
      title: ApiResponse[TokenResponse]
...
    FirstAdminSignupRequest:
      properties:
        name:
          type: string
          maxLength: 40
          minLength: 3
          title: Name
          description: Admin's full name
        email:
          type: string
          format: email
          title: Email
          description: Admin's email address
        password:
          type: string
          minLength: 8
          title: Password
          description: Admin's password
      type: object
      required:
      - name
      - email
      - password
      title: FirstAdminSignupRequest
      description: Request schema for first-time admin account creation.
    FirstAdminSignupResponse:
      properties:
        user_id:
          type: integer
          title: User Id
          description: ID of the newly created admin user
        name:
          type: string
          title: Name
          description: Admin's name
        email:
          type: string
          format: email
          title: Email
          description: Admin's email
      type: object
      required:
      - user_id
      - name
      - email
      title: FirstAdminSignupResponse
      description: Response schema for successful first-time admin signup.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
          description: User's email address
        password:
          type: string
          minLength: 1
          title: Password
          description: User's password
      additionalProperties: false
      type: object
      required:
      - email
      - password
      title: LoginRequest
...
    RefreshRequest:
      properties:
        refresh_token:
          type: string
          minLength: 1
          title: Refresh Token
          description: Valid refresh token
      additionalProperties: false
      type: object
      required:
      - refresh_token
      title: RefreshRequest
    TemplateCreate:
      properties:
        name:
          type: string
          minLength: 1
          title: Name
        description:
          type: string
          title: Description
          default: ''
        category:
          type: string
          title: Category
          default: general
        workflow_data:
          additionalProperties: true
          type: object
          title: Workflow Data
        is_public:
          type: boolean
          title: Is Public
          default: false
      type: object
      required:
      - name
      - workflow_data
      title: TemplateCreate
      description: Schema for creating a new template.
    TemplateDetail:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        category:
          type: string
          title: Category
        workflow_data:
          additionalProperties: true
          type: object
          title: Workflow Data
        usage_count:
          type: integer
          title: Usage Count
        is_public:
          type: boolean
          title: Is Public
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        created_by:
          anyOf:
          - type: integer
          - type: 'null'
          title: Created By
      type: object
      required:
      - id
      - name
      - description
      - category
      - workflow_data
      - usage_count
      - is_public
      - created_at
      - updated_at
      - created_by
      title: TemplateDetail
      description: Detailed template information including workflow data.
    TemplateSummary:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        category:
          type: string
          title: Category
        usage_count:
          type: integer
          title: Usage Count
        is_public:
          type: boolean
          title: Is Public
      type: object
      required:
      - id
      - name
      - description
      - category
      - usage_count
      - is_public
      title: TemplateSummary
      description: Template summary for listing templates.
    TemplateWorkflowData:
      properties:
        workflow_data:
          additionalProperties: true
          type: object
          title: Workflow Data
      type: object
      required:
      - workflow_data
      title: TemplateWorkflowData
      description: Schema for template workflow data response.
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
          description: JWT access token
        refresh_token:
          type: string
          title: Refresh Token
          description: Refresh token for obtaining new access tokens
        token_type:
          type: string
          title: Token Type
          description: Token type, always 'bearer'
          default: bearer
        expires_in:
          type: integer
          title: Expires In
          description: Access token expiration time in seconds
      type: object
      required:
      - access_token
      - refresh_token
      - expires_in
      title: TokenResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth/login